Work in progress to have a a draw command package for android

+++ platforms/android/simple-guido-editor/src/fr/grame/simpleguidoeditor/drawcommand/AndroidCanvasDrawStringCommand.java
@@ -16,14 +16,13 @@ public class AndroidCanvasDrawStringCommand extends DrawStringCommand implements
   public void drawToCanvas(Canvas canvas, GuidoCanvasView view) {
     double x = _x;
     double y = _y;
-    //double size = view._MUSIC_FONT_SIZE;
 
     view.correctTransformMatrix(canvas);
 
     Paint paint = new Paint();
     paint.setTypeface(view._TEXT_FONT);
     paint.setColor(view._FONT_COLOR);
-    paint.setTextSize((float)(view._TEXT_FONT_SIZE));
+    paint.setTextSize(view._TEXT_FONT_SIZE);
 
     canvas.drawText(_s, (float) x, (float) y, paint);
     view.resetTransformMatrix(canvas);
